updating mock scaffolder

+++ sayedha-install-scaffold-all.ps1
@@ -1,8 +1,6 @@
 [cmdletbinding()]
 param(
     [string]$srcDir = "$PSScriptRoot",
-    [string]$dotnetRoot="$PSScriptRoot/.dotnet",
-    [string]$dotnetRootX86="$PSScriptRoot/.dotnet/x86",
     [string]$dotnetMultiLevelLookup=0,
     [string]$version="9.0.1-dev",
     [string]$defaultNupkgDir="$env:USERPROFILE\.nuget\packages",
@@ -12,24 +10,28 @@ param(
     [string]$nupkgDir="$PSScriptRoot/artifacts/packages/Debug/Shipping/"
 )
 
-function Clean(){
+function Clean{
     Write-Host  "### Clean" -ForegroundColor Red
-    DeleteFolder -f $artifactsDir
+    DeleteFolder -folder $artifactsDir
 }
-function Build(){
+function Build{
     Write-Host "### Build" -ForegroundColor Red
     # powershell -ExecutionPolicy ByPass -NoProfile -command "& """%~dp0eng\common\Build.ps1""" -build -restore -pack %*"
-    & $srcDir/eng/common/Build.ps1 -build -restore -pack -bl
+    & $srcDir/eng/common/Build.ps1 -build -restore -pack -bl /clp:ErrorsOnly
 }
 
-function Install(){
+function Install{
     Write-Host "### Install" -ForegroundColor Red
+    if (-not (Test-Path $nupkgDir)) {
+        Write-Error "Package source directory not found: '$nupkgDir'"
+        return
+    }
     & dotnet tool install -g Microsoft.dotnet-scaffold --add-source $nupkgDir --version $version
     & dotnet tool install -g Microsoft.dotnet-scaffold-aspire --add-source $nupkgDir --version $version
     & dotnet tool install -g Microsoft.dotnet-scaffold-aspnet --add-source $nupkgDir --version $version
 }
 
-function Uninstall(){
+function Uninstall{
     Write-Host "### Uninstall" -ForegroundColor Red
     & dotnet tool uninstall -g Microsoft.dotnet-scaffold
     & dotnet tool uninstall -g Microsoft.dotnet-scaffold-aspire
@@ -87,23 +89,26 @@ function DeleteFile{
     }
 }
 
-function Initalize(){
-    Write-Host "### Initalize" -ForegroundColor Red
-    if ($env:PATH -notmatch [regex]::Escape($DOTNET_ROOT)) {
-        $env:PATH = "$DOTNET_ROOT;$env:PATH"
+function KillDotnet{
+    Write-Host "### KillDotnet" -ForegroundColor Red
+    # call taskkill /f /im dotnet.exe
+    $processes = Get-Process -Name dotnet.exe -ErrorAction SilentlyContinue
+    if($processes){
+        foreach($p in $processes){
+            Stop-Process -Id $p.Id -Force
+        }
     }
-    $env:DOTNET_MULTILEVEL_LOOKUP = $dotnetMultiLevelLookup
-    $env:DOTNET_ROOT=$dotnetRoot
 }
-function ResetEnv(){
+
+function ResetEnv{
     Write-Host "### ResetEnv" -ForegroundColor Red
-    $env:PATH = ($env:PATH -split ';' | Where-Object { $_ -ne $DOTNET_ROOT }) -join ';'
     $env:DOTNET_MULTILEVEL_LOOKUP = ""
     $env:DOTNET_ROOT=""
     $env:DOTNET_INSTALL_DIR=""
 }
 
 ### Start script
+KillDotnet
 Clean
 Build
 ResetEnv
